Python-serice/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
python-dotenv
email-validator
pymongo